+++ source/extensions/isaacsim.app.setup/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.11.0] - 2026-03-23
+### Added
+- Emit telemetry event for app startup duration via isaacsim.core.telemetry
+### Changed
+- Updated return types
+
 ## [1.10.0] - 2026-03-04
 ### Changed
 - Add Overview.md, python_api.md, SETTINGS.md and update docstrings